Introduce remoteId field to distinguish synced objects.
remoteId tries to decouple the concept of synced resource from the id
field. In some cases, the id is not a value decided by the database (an
autoincremental number), but an attribute of the object. In this
situation, checking if the id is set or not is not a valid strategy to
distinguish synced objects. remoteId is a field that is only set after
db syncs, marking an object as sync or not. urls are built using the new
remoteId so that id attributes can be updated without breaking the
resource url.